REF: Move `CrashDiagnostics` into its own module
